Seperated style into a different file and changed some the login buttonto touchable opacity
